Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
Werkzeug==3.0.1

# Optional: For better OCR (uncomment if needed)
# pytesseract==0.3.10
# google-cloud-vision==3.5.0

